% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addinslist-package.R
\docType{package}
\name{addinslist-package}
\alias{addinslist}
\alias{addinslist-package}
\title{addinslist: Discover and Install Useful RStudio Addins}
\description{
Browse through a continuously updated list of existing RStudio addins and install/uninstall their corresponding packages.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/daattali/addinslist}
  \item Report bugs at \url{https://github.com/daattali/addinslist/issues}
}

}
\author{
\strong{Maintainer}: Dean Attali \email{daattali@gmail.com} (\href{https://orcid.org/0000-0002-5645-3493}{ORCID})

}
\keyword{internal}
